Extract dashboard logic
